Integrate ML predictions, add favicons, add about page

+++ website/models/next_delay.py
@@ -3,7 +3,7 @@ from website.models import Trips
 
 class NextDelay(models.Model):
     trip = models.ForeignKey(Trips, on_delete=models.CASCADE)
-    delay = models.FloatField()
+    delay = models.IntegerField()
 
     def __str__(self):
         return self.delay
